+++ lib/outliers/exceptions.rb
@@ -27,6 +27,9 @@ module Outliers
     class UnknownVerification < Base
     end
 
+    class UnknownFilter < Base
+    end
+
     class UnknownProvider < Base
     end