+++ lib/outliers/resources/aws/ec2/image.rb
@@ -1,0 +1,13 @@
+module Outliers
+  module Resources
+    module Aws
+      module Ec2
+        class Image < Resource
+          def self.key
+            'image_id'
+          end
+        end
+      end
+    end
+  end
+end