+++ lib/outliers/resources/github.rb
@@ -1,2 +1,0 @@
-require 'outliers/resources/github/repo'
-require 'outliers/resources/github/repo_collection'